# Cache heap::downheap() root comparison (optimize heap cmp call)